+++ CHANGELOG.md
@@ -6,7 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Autoplay events
+- Full screen events
+- Game loading event
+- Open Home event
 
+### Changed
+- gameLoaded now sends `status: "done"`
+- gameLoadError sends `reason` instead of `error`
 ## 0.1.0 - 2022-12-01
 ### Added
 - Support more events on `ValkyrieWrapper`